Extract related query to its own module.
