+++ lib/re/listings/listings.ex
@@ -18,9 +18,13 @@ defmodule Re.Listings do
   alias Ecto.Changeset
 
   defdelegate authorize(action, user, params), to: Re.Listings.Policy
+  defdelegate related(listing, limit), to: Re.Listings.Related
+  defdelegate related(listing), to: Re.Listings.Related
 
   @active_listings_query from(l in Listing, where: l.is_active == true)
   @order_by_position from(i in Image, where: i.is_active == true, order_by: i.position)
+  def active_listings_query, do: @active_listings_query
+  def order_by_position, do: @order_by_position
 
   def paginated(params) do
     @active_listings_query
@@ -88,48 +92,6 @@ defmodule Re.Listings do
     |> Repo.all()
     |> Enum.map(&Map.get(&1, :listing))
     |> check_if_exists()
-  end
-
-  def related(listing, limit \\ :no_limit) do
-    ~w(price address)a
-    |> do_related(listing, @active_listings_query)
-    |> Enum.reject(fn %{id: id} -> id == listing.id end)
-    |> Enum.uniq_by(fn %{id: id} -> id end)
-    |> limit_results(limit)
-    |> Repo.preload([:address, images: @order_by_position])
-    |> okd()
-  end
-
-  defp limit_results(result, :no_limit), do: result
-  defp limit_results(result, limit), do: Enum.take(result, limit)
-
-  defp okd(arg), do: {:ok, arg}
-
-  defp do_related([], _, _), do: featured()
-
-  defp do_related([_attr | rest] = attrs, listing, query) do
-    listing
-    |> Repo.preload(:address)
-    |> Map.take(attrs)
-    |> Enum.reduce(query, &build_query(&1, &2))
-    |> Repo.all()
-    |> Enum.concat(do_related(rest, listing, query))
-  end
-
-  defp build_query({:address, address}, query) do
-    from(
-      l in query,
-      join: a in assoc(l, :address),
-      where: ^address.neighborhood == a.neighborhood
-    )
-  end
-
-  defp build_query({:price, price}, query) do
-    price_diff = price * 0.25
-    floor = trunc(price - price_diff)
-    ceiling = trunc(price + price_diff)
-
-    from(l in query, where: l.price >= ^floor and l.price <= ^ceiling)
   end
 
   @top_4_listings_query from(l in Listing, where: l.is_active == true, order_by: [desc: l.score])
